commit test 18

+++ src/module_reproduction.f90
@@ -46,7 +46,7 @@ contains
 
             npp_rep = nppa * 0.04  ! 4% do NPP disponível para reprodução
              ! Calculando o número de sementes produzidas
-            n_seed = int(npp_rep / seed_mass)
+            n_seed = nint(npp_rep / seed_mass)
             print *, "n_seed:", n_seed
             npp_after = nppa - npp_rep !remaining npp
         else
